try to fix XML::Simple problems with parse_string on files
Still can't reproduce this locally (even with XML::Simple 2.24).

In some cases, XML::Simple->parse_string() seems to be called with
a file and not a string, resulting in this error message:

   Nonblocking reads only make sense if you're giving me file handles,
   y'know (well, actually sockets). Use parse_file at…

So use XML::Simple->xml_in() which should be able to call either
parse_file() or parse_string() as needed.

This could fix issue #7.